app/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
drf-yasg==1.21.4
Pillow==9.4.0
psycopg2==2.9.5
pytest==7.2.1
pytest-django==4.5.2
